% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_partialOmegaSq.R
\name{adonis_partialOmegaSq}
\alias{adonis_partialOmegaSq}
\title{Calculate partial Omega-squared (effect-size calculation) for PERMANOVA and add it to the input object}
\usage{
adonis_partialOmegaSq(adonisOutput)
}
\arguments{
\item{adonisOutput}{An adonis object}
}
\value{
Original adonis object with the partial Omega-squared values added
}
\description{
Calculate partial Omega-squared (effect-size calculation) for PERMANOVA and add it to the input object
}
