% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_scientific.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Fancy Scientific Notation}
\usage{
fancy_scientific(l)
}
\arguments{
\item{l}{Numberic label input}
}
\value{
Value in scientific notation
}
\description{
Code from Brian Diggs \url{https://groups.google.com/forum/#!topic/ggplot2/a_xhMoQyxZ4}
}
