% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2bayes.R
\name{p2bayes}
\alias{p2bayes}
\title{Estimate maximum possible Bayes factor from p-value}
\usage{
p2bayes(x)
}
\arguments{
\item{x}{p-value}
}
\value{
Maximum possible Bayes factor
}
\description{
Estimate maximum possible Bayes factor from p-value
}
