% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{predictdf.TLS}
\alias{predictdf.TLS}
\title{Internal prediction functions for tls smooth}
\usage{
predictdf.TLS(model, xseq, se, level)
}
\arguments{
\item{model}{Input model}

\item{xseq}{x-values used for prediction}

\item{se}{Predict error or not}

\item{level}{Confidence level}
}
\description{
Internal prediction functions for tls smooth
}
