% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_by_group.R
\name{prune_by_group}
\alias{prune_by_group}
\title{Prune taxa in a phyloseq object by group}
\usage{
prune_by_group(phy, group, P = 0.5)
}
\arguments{
\item{phy}{phyloseq object}

\item{group}{Variable in phyloseq object to group by}

\item{P}{Proportion taxa has to be present in, in each group}
}
\value{
Filtered phyloseq object
}
\description{
Filter taxa to only include taxa that are present in a certain proportion in each group defined by a variable in sample_data()
}
