% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_elisa.R
\name{read_elisa}
\alias{read_elisa}
\title{Load data from a ELISA plate reader}
\usage{
read_elisa(paths, descriptions = 0)
}
\arguments{
\item{paths}{The path(s) for the excel file(s)}

\item{descriptions}{Number of descriptive sheets. Default 0, i.e. only OD measurements are loaded}
}
\value{
A dataframe with all data. Descriptions are called V1, V2, and so on, in the same order as in the excel sheets.
}
\description{
Load excel sheet data from ELISA plate reader (96 Well plates), and add on descriptions of what is in the wells, and output a neat data.frame.
The excel file(s) should contain a sheet with OD measurements, and can contain additional sheets with descriptions of what is in the wells, e.g. WT, Contol and so on.
The measurements and desciptive sheets should be set up as a 96 Well plate with including A to H and 1 to 12.
}
