% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.R
\name{tls}
\alias{tls}
\title{Total Least Squares for ggplot stat_smooth and geom_smooth
With bootstrapped confidence intervals}
\usage{
tls(formula, data, ...)
}
\arguments{
\item{formula}{formula for the tls}

\item{data}{Input data.frame}

\item{...}{Unused argument}
}
\description{
Total Least Squares for ggplot stat_smooth and geom_smooth
With bootstrapped confidence intervals
}
